add some scripts
